Crate pallet_finality_tracker

FRAME Pallet that tracks the last finalized block, as perceived by block authors.

Structs

GenesisConfig
Genesis config for the module, allow to build genesis storage.

Module

Enums

Call
Dispatchable calls.

Error

Constants

DEFAULT_REPORT_LATENCY

DEFAULT_WINDOW_SIZE

Traits

OnFinalizationStalled
Called when finalization stalled at a given number.

Trait