Trait pallet_finality_tracker::Trait

pub trait Trait: SystemTrait {
    type OnFinalizationStalled: OnFinalizationStalled<Self::BlockNumber>;
    type WindowSize: Get<Self::BlockNumber>;
    type ReportLatency: Get<Self::BlockNumber>;
}

Associated Types

type OnFinalizationStalled: OnFinalizationStalled<Self::BlockNumber>

Something which can be notified when the timestamp is set. Set this to () if not needed.

type WindowSize: Get<Self::BlockNumber>

The number of recent samples to keep from this chain. Default is 101.

type ReportLatency: Get<Self::BlockNumber>

The delay after which point things become suspicious. Default is 1000.

Loading content...